refactor: 의존성 순환 수정
